Fix lost genotype when deleting 1st group
A bug was identified that showed that when the group being deleted with
a call to delete.group was the first one stored in the interal array,
one of the genotypes in the second group would be overwritten with NULLs.
This was a bug in the condense_allele_matrix function that is now fixed.