Allow integer vector inputs to be of numeric type
R has two data types for numbers: numerics and integers. Internally genomicSimulation prefers integers, but it would be more user-friendly if users could pass parameters to these functions without needing to wrap them in `as.integer()` or needing to append L to the end of all numeric constants.

`see.group.data`, `delete.group`, and `combine.groups` no longer raise an error when the group numbers passed are of numeric type instead of integer type. Instead, as long as the numbers are whole numbers, they perform the type conversion. Likewise, `make.targeted.crosses` no longer raises an error when it is passed parent indexes that are of numeric type.